requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
qdrant-client==1.7.0
openai==1.3.5
numpy==1.24.3
tiktoken==0.5.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
better-exceptions==0.3.3
python-json-logger==2.0.7
asyncpg==0.29.0
opencv-python==4.8.1.78
face-recognition==1.3.0
imutils==0.5.4